Smart home using ATMEGA32
Interfacing with AVR (ATMEGA32) to interact with HW like (Temperature sensor, smoke, IR, LCD, Keypad, Bluetooth module)  with GUI to transfer the data from and to the user.

The algorithm contains password checking before entering home using LCD& Keypad , the algorithm waits for the door to be opened(using IR sensor)
to start receiving data from the user.

BLE that transfers data from the user to change the home mode to
1-manual:
The user can manually turn on/off hardware like lamp or fan
 2-Autonomeus
The home will turn on/off the fan and lamp based on the readings of the temperature sensor to maintain the room temperature at the acceptable range .